fix: Fix request controller always waiting for all nodes.
Previously, the controller waited for all nodes to respond, even when one already confirmed the check. This would lead to an additional second in processing if one node was always busy. Furthermore, this fixes an issue where one node might always be unavailable, leading to 503 error messages due to timeouts.